% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolDoveR.R
\name{cool2IntSet}
\alias{cool2IntSet}
\title{Read .cool/.mcool sparse matrix into a GInteractions}
\usage{
cool2IntSet(files, res = NULL, cores = detectCores())
}
\arguments{
\item{files}{A character vector of paths to HDF5 stored cool (uni-dimension)
or mcool (multi-dimension sparse matrix). If using a multi-dimesion,
the resolution of one of the dimension need to be passed to res.}

\item{res}{'NULL' if using a uni-dimensional cool file or the resolution of one of layer in the mcool file}

\item{cores}{An integer for the number of parallel thread to convert the file}
}
\value{
An InteractionSet of all common pairs with the number of interaction for each pairs
}
\description{
Read .cool/.mcool sparse matrix into a GInteractions
}
