% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolR.R
\name{getBins}
\alias{getBins}
\title{Accessing the genomic bins in a cool/mcool file}
\usage{
getBins(file, res)
}
\arguments{
\item{file}{Path to a HDF5 stored cool (uni-dimension) or mcool (multi-dimension sparse matrix).
If using a multi-dimesion, the resolution of one of the dimension need to be passed to res}

\item{res}{'NULL' if using a uni-dimensional cool file or the resolution of one of layer in the mcool file}
}
\description{
Retrieves the chromosome bins forming from a HDF5 stored uni- or multi-dimentional
contact matrix (cool or mcool file) into a seqinfo object
}
\examples{

}
