% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolDoveR.R
\name{read.cool}
\alias{read.cool}
\title{Read .cool/.mcool sparse matrix into a GInteractions}
\usage{
read.cool(
  file,
  res = NULL,
  chr1 = NULL,
  start1 = NULL,
  end1 = NULL,
  chr2 = NULL,
  start2 = NULL,
  end2 = NULL
)
}
\arguments{
\item{file}{Path to a HDF5 stored cool (uni-dimension) or mcool (multi-dimension sparse matrix).
If using a multi-dimesion, the resolution of one of the dimension need to be passed to res}

\item{res}{'NULL' if using a uni-dimensional cool file or the resolution of one of layer in the mcool file}

\item{chr1}{'NULL' or the chromosome name of the first pairs to return}

\item{start1}{'NULL' or the start of a region within chr1 to return the pairs from.
If chr1 is set and start1 is NULL this will be set to 1 by default}

\item{end1}{'NULL' or the end of a region within chr1 to return the pairs from.
If chr1 is set and end1 is NULL this will be set to the lenght of chr1  by default}

\item{chr2}{'NULL' or the chromosome name of the first pairs to return}

\item{start2}{'NULL' or the start of a region within chr2 to return the pairs from.
If chr1 is set and start1 is NULL this will be set to 1 by default}

\item{end2}{'NULL' or the end of a region within chr1 to return the pairs from.
If chr2 is set and end1 is NULL this will be set to the lenght of chr2  by default}
}
\value{
A GInteraction object of the contact matrix dataset
}
\description{
This function reads a compressed sparse row (CSR) storage scheme for a matrix
created by the cooler application (https://github.com/mirnylab/cooler) stored in a
HDF5 data storage. This utility can read the matrix or only specfic region of the genome
eihter from one pair or a pair of region. If using a multi-dimension dataset (ie .mcool), the resolution
needed need to be specified.
}
